requirements.txt
py-cord @ git+https://github.com/Pycord-Development/pycord@master
vosk==0.3.45
numpy==1.26.4
PyNaCl==1.5.0
python-dotenv==1.0.1
